+++ stdlib/lib.core.d.ts
@@ -1608,6 +1608,7 @@ interface FSFileHandle {
     write(buffer: Buffer, offset?: number, length?: number, position?: number | null, ...ignored: any[]): Promise<FSFileWriteResult>;
     readv(buffers: Buffer[], position?: number | null, ...ignored: any[]): Promise<FSFileReadvResult>;
     writev(buffers: Buffer[], position?: number | null, ...ignored: any[]): Promise<FSFileWritevResult>;
+    appendFile(data: string | Buffer, options?: FSAppendFileEncodingOptions, ...ignored: any[]): Promise<void>;
     chmod(mode: number, ...ignored: any[]): Promise<void>;
     chown(uid: number, gid: number, ...ignored: any[]): Promise<void>;
     utimes(atime: FSFileTime, mtime: FSFileTime, ...ignored: any[]): Promise<void>;